+++ Release Notes.md
@@ -10,7 +10,9 @@
 ### 🗑 Deprecations
 
 * `Presentation/Toast` has been deprecated. Use the new [SystemNotification](https://github.com/danielsaidi/SystemNotification) library instead.
-
+* `View+Alert` `alert(context:)` has been renamed to `alert(_ :)`.
+* `View+FullScreenCover` `fullScreenCover(context:)` has been renamed to `fullScreenCover(_ :)`.
+* `View+Sheet` `sheet(context:)` has been renamed to `sheet(_ :)`.
 
 
 ## 2.3.3